Rename Field with Line and field prop with label.
