+++ app/components/Company/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/styles';
 import Rating from 'material-ui-rating';
-import Field from 'components/Field/Loadable';
+import Line from 'components/Line/Loadable';
 
 const useStyles = makeStyles({
   root: {
@@ -17,12 +17,12 @@ const Company = props => {
 
   return (
     <div className={classes.root}>
-      <Field field="Name">{company.name}</Field>
-      <Field field="Address">{company.address}</Field>
-      <Field field="Business">{company.business}</Field>
-      <Field field="Rating">
+      <Line label="Name">{company.name}</Line>
+      <Line label="Address">{company.address}</Line>
+      <Line label="Business">{company.business}</Line>
+      <Line label="Rating">
         <Rating value={company.rating} max={10} />
-      </Field>
+      </Line>
     </div>
   );
 };
