+++ app/components/Line/index.js
@@ -3,26 +3,26 @@ import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/styles';
 
 const useStyles = makeStyles({
-  field: {
+  label: {
     float: 'left',
     width: '50%',
   },
 });
 
 const Field = props => {
-  const { field, children } = props;
+  const { label, children } = props;
   const classes = useStyles();
 
   return (
     <div>
-      <div className={classes.field}>{field}</div>
-      <div className={classes.field}>{children}</div>
+      <div className={classes.label}>{label}</div>
+      <div className={classes.label}>{children}</div>
     </div>
   );
 };
 
 Field.propTypes = {
-  field: PropTypes.string.isRequired,
+  label: PropTypes.string.isRequired,
   children: PropTypes.oneOfType([PropTypes.node, PropTypes.element]),
 };
